requirements.txt
flask
gunicorn
pytest
pycodestyle
flask_cors
google-auth-httplib2
google-auth-oauthlib
pymongo[srv]
dnspython
dotenv
